+++ aws-dotnet-rest-api-with-dynamodb/DotNetServerless.sln
@@ -9,9 +9,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{DBC29D13
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetServerless.Domain", "src\DotNetServerless.Domain\DotNetServerless.Domain.csproj", "{EE96C435-FFE9-4F98-9865-47B0DC7C2427}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetServerless.Functions.Tests", "tests\DotNetServerless.Functions.Tests\DotNetServerless.Functions.Tests.csproj", "{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetServerless.Lambda", "src\DotNetServerless.Lambda\DotNetServerless.Lambda.csproj", "{9A74B45A-50B6-44C8-A2D7-F7778A61F18F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetServerless.Tests", "tests\DotNetServerless.Tests\DotNetServerless.Tests.csproj", "{045DC8E9-8CA6-4B76-9C4A-781AB9589700}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,18 +38,6 @@ Global
 		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|x64.Build.0 = Release|Any CPU
 		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|x86.ActiveCfg = Release|Any CPU
 		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|x86.Build.0 = Release|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x64.Build.0 = Debug|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x86.Build.0 = Debug|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x64.ActiveCfg = Release|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x64.Build.0 = Release|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x86.ActiveCfg = Release|Any CPU
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x86.Build.0 = Release|Any CPU
 		{9A74B45A-50B6-44C8-A2D7-F7778A61F18F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9A74B45A-50B6-44C8-A2D7-F7778A61F18F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9A74B45A-50B6-44C8-A2D7-F7778A61F18F}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -62,10 +50,22 @@ Global
 		{9A74B45A-50B6-44C8-A2D7-F7778A61F18F}.Release|x64.Build.0 = Release|Any CPU
 		{9A74B45A-50B6-44C8-A2D7-F7778A61F18F}.Release|x86.ActiveCfg = Release|Any CPU
 		{9A74B45A-50B6-44C8-A2D7-F7778A61F18F}.Release|x86.Build.0 = Release|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Debug|x64.Build.0 = Debug|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Debug|x86.Build.0 = Debug|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Release|Any CPU.Build.0 = Release|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Release|x64.ActiveCfg = Release|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Release|x64.Build.0 = Release|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Release|x86.ActiveCfg = Release|Any CPU
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{EE96C435-FFE9-4F98-9865-47B0DC7C2427} = {E45346EE-74B7-4F5B-943C-FEFDE57124D0}
-		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2} = {DBC29D13-84FE-4A90-B785-E325BDD494A8}
 		{9A74B45A-50B6-44C8-A2D7-F7778A61F18F} = {E45346EE-74B7-4F5B-943C-FEFDE57124D0}
+		{045DC8E9-8CA6-4B76-9C4A-781AB9589700} = {DBC29D13-84FE-4A90-B785-E325BDD494A8}
 	EndGlobalSection
 EndGlobal
