+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Lambda/Program.cs
@@ -1,4 +1,4 @@
-﻿namespace DotNetServerless.Functions
+﻿namespace DotNetServerless.Lambda
 {
   public class Program
   {
