+++ aws-dotnet-rest-api-with-dynamodb/tests/DotNetServerless.Tests/Functions/CreateItemFunctionTests.cs
@@ -1,43 +1,42 @@
-using System;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Amazon.Lambda.APIGatewayEvents;
 using DotNetServerless.Domain.Entity;
 using DotNetServerless.Domain.Infrastructure.Repositories;
 using DotNetServerless.Domain.Requests;
-using DotNetServerless.Functions.Lambda;
+using DotNetServerless.Lambda;
+using DotNetServerless.Lambda.Functions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Moq;
 using Newtonsoft.Json;
 using Xunit;
 
-namespace DotNetServerless.Functions.Tests.Lambda
+namespace DotNetServerless.Tests.Functions
 {
-  public class GetItemFunctionTests
+  public class CreateItemFunctionTests
   {
-    public GetItemFunctionTests()
+    public CreateItemFunctionTests()
     {
       _mockRepository = new Mock<IItemRepository>();
-      _mockRepository.Setup(_ => _.GetById<Item>(It.IsAny<string>(), It.IsAny<CancellationToken>())).ReturnsAsync(new List<Item>{ new Item{ Id = Guid.NewGuid()}});
+      _mockRepository.Setup(_ => _.Save(It.IsAny<Item>(), It.IsAny<CancellationToken>())).Returns(Task.CompletedTask);
 
       var serviceCollection = Startup.BuildContainer();
 
       serviceCollection.Replace(new ServiceDescriptor(typeof(IItemRepository), _ => _mockRepository.Object,
         ServiceLifetime.Transient));
 
-      _sut = new GetItemFunction(serviceCollection.BuildServiceProvider());
+      _sut = new CreateItemFunction(serviceCollection.BuildServiceProvider());
     }
 
-    private readonly GetItemFunction _sut;
+    private readonly CreateItemFunction _sut;
     private readonly Mock<IItemRepository> _mockRepository;
 
     [Fact]
     public async Task run_should_trigger_mediator_handler_and_repository()
     {
-      await _sut.Run(new APIGatewayProxyRequest{ Body = JsonConvert.SerializeObject(new GetItemRequest())});
-      _mockRepository.Verify(_ => _.GetById<Item>(It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Once);
+      await _sut.Run(new APIGatewayProxyRequest {Body = JsonConvert.SerializeObject(new CreateItemRequest())});
+      _mockRepository.Verify(_ => _.Save(It.IsAny<Item>(), It.IsAny<CancellationToken>()), Times.Once);
     }
   }
 }
